Rename Offlinedata class to Data and and add todolist to Data
