--- src/main/java/com/rettichlp/UnicacityAddon/base/json/todo/Todolist.java
@@ -1,12 +1,0 @@
-package com.rettichlp.UnicacityAddon.base.json.todo;
-
-import java.util.List;
-
-public class Todolist {
-
-    List<TodolistEntry> todolistEntryList;
-
-    public List<TodolistEntry> getTodolistEntryList() {
-        return todolistEntryList;
-    }
-}
